GoogleCloudPlatform/covid19-rapid-response-demo/twilio-sms/requirements.txt
twilio
flask
dialogflow
gunicorn
